requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-multipart==0.0.9
pillow==10.3.0
requests==2.31.0
pandas==2.2.2
google-cloud-vision==3.7.3
google-auth==2.29.0
